+++ test/matchers/document_test.rb
@@ -5,4 +5,5 @@ describe "Document matchers" do
 
   it { must be_document }
   it { must be_paranoid }
+  it { must be_versioned }
 end