+++ test/models/person.rb
@@ -1,6 +1,7 @@
 class Person
   include Mongoid::Document
   include Mongoid::Paranoia
+  include Mongoid::Versioning
 
   field :name, type: String, default: "me"
 end